[ex2 files] practical q1

+++ ex2.py
@@ -41,7 +41,7 @@ def get_args():
     parser.add_argument('--image_path', default="", help='The path to keep the learned image')
     parser.add_argument('--reg_factor', type=float, default=0.2, help='The regression (lambda) value')
     parser.add_argument('--min_target_activation', type=float, default=400, help='The min value for the neuron activation')
-    parser.add_argument('--max_iter', type=float, default=5000, help='The maximum iterations')
+    parser.add_argument('--max_iter', type=int, default=5000, help='The maximum iterations')
     parser.add_argument('--neuron_layer_idx', "-nl", type=int, default=21, help='The index of the require neuron')
     parser.add_argument('-ni', '--neuron_idx_list', type=int, default=[0], action='append', help='The indices of the neuron (-ni=1, -ni=2 -ni=3)')
